requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.4
django-ckeditor==5.7.1
django-jet==1.0.8
django-js-asset==1.2.2
djangorestframework==3.10.2
facebook-sdk==3.0.0
idna==2.8
oauthlib==3.1.0
Pillow==6.1.0
PySocks==1.7.1
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
tweepy==3.8.0
urllib3==1.25.6
